Add a image header.
I really like this program, but I think that it would add a lot to add a basic screenshot at the top of the screen so that people know what they are downloading. I know that this image is on KDE's replicator site, but I thought it would fit in nicely on the GitHub page.